+++ src/renderer/src/tools-library/developer-collective-store.ts
@@ -1,0 +1,18 @@
+import { manualReloadStore } from '@holochain-open-dev/stores';
+import { ActionHash } from '@holochain/client';
+import { ToolsLibraryClient } from './tools-library-client.js';
+
+export class DeveloperCollectiveStore {
+  developerCollectiveHash: ActionHash;
+
+  constructor(
+    public toolsLibraryClient: ToolsLibraryClient,
+    developerCollectiveHash: ActionHash,
+  ) {
+    this.developerCollectiveHash = developerCollectiveHash;
+  }
+
+  allTools = manualReloadStore(async () =>
+    this.toolsLibraryClient.getToolsForDeveloperCollective(this.developerCollectiveHash),
+  );
+}
